CCm-8308: implement metadata correctly
