+++ src/app/signout/page.tsx
@@ -1,23 +1,20 @@
-'use client';
-
 import React from 'react';
 import { SignOut } from '@/src/components/molecules/SignOut/SignOut';
 import content from '@/src/content/content';
-import { MetaTitle } from '@/src/components/atoms/MetaTitle/MetaTitle';
+import { Metadata } from 'next';
+
+const {
+  pages: { signOutPage },
+} = content;
+
+export const metadata: Metadata = {
+  ...signOutPage.meta,
+};
 
 export default function Page() {
-  const {
-    pages: { signOutPage },
-  } = content;
   return (
-    <>
-      <MetaTitle
-        title={signOutPage.meta.title}
-        description={signOutPage.meta.description}
-      />
-      <SignOut>
-        <p>{signOutPage.content}</p>
-      </SignOut>
-    </>
+    <SignOut>
+      <p>{signOutPage.content}</p>
+    </SignOut>
   );
 }
