--- src/components/atoms/MetaTitle/MetaTitle.tsx
@@ -1,11 +1,0 @@
-type Props = {
-  title: string;
-  description: string;
-};
-
-export const MetaTitle = ({ title, description }: Props) => (
-  <>
-    <title>{title}</title>
-    <meta name='description' content={description} />
-  </>
-);
